crs/libs/libCRS/requirements.txt
coloredlogs
pyyaml
GitPython
requests==2.32.3
tabulate==0.9.0
gspread
google-auth
pandas
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
loguru
